Funziona senza il Magnetometro ma quandi si inseriscono le letture del magnetometro "da i numeri"

+++ FIMU_ITG3200.cpp
@@ -38,18 +38,14 @@ using namespace std;
 
 
 ITG3200::ITG3200() {
-    setGains(1.0,1.0,1.0);
-    setOffsets(0.0,0.0,0.0);
-    setRevPolarity(0,0,0);
-    //Wire.begin();       //Normally this code is called from setup() at user code
-    //but some people reported that joining I2C bus earlier
-    //apparently solved problems with master/slave conditions.
-    //Uncomment if needed.
+    setGains(1.0, 1.0, 1.0);
+    setOffsets(0.0, 0.0, 0.0);
+    setRevPolarity(0, 0, 0);
 }
 
 
 void
-ITG3200::init(unsigned int  address) {
+ITG3200::init(uint16_t  address) {
     // Uncomment or change your default ITG3200 initialization
 
     // fast sample rate - divisor = 0 filter = 0 clocksrc = 0, 1, 2, or 3  (raw values)
@@ -67,7 +63,7 @@ ITG3200::init(unsigned int  address) {
 
 
 void
-ITG3200::init(unsigned int address, byte _SRateDiv, byte _Range, byte _filterBW, byte _ClockSrc, bool _ITGReady, bool _INTRawDataReady) {
+ITG3200::init(uint16_t address, byte _SRateDiv, byte _Range, byte _filterBW, byte _ClockSrc, bool _ITGReady, bool _INTRawDataReady) {
     _dev_address = address;
     setSampleRateDiv(_SRateDiv);
     setFSRange(_Range);
@@ -81,14 +77,14 @@ ITG3200::init(unsigned int address, byte _SRateDiv, byte _Range, byte _filterBW,
 
 byte
 ITG3200::getDevAddr() {
-    /*readmem(WHO_AM_I, 1, &_buff[0]);
-  return _buff[0];  */
+  /*readmem(WHO_AM_I, 1, &_buff[0]);
+    return _buff[0];  */
     return _dev_address;
 }
 
 
 void
-ITG3200::setDevAddr(unsigned int  _addr) {
+ITG3200::setDevAddr(uint16_t  _addr) {
     writemem(WHO_AM_I, _addr);
     _dev_address = _addr;
 }
@@ -235,13 +231,13 @@ ITG3200::isRawDataReady() {
 
 void
 ITG3200::readTemp(float *_Temp) {
-    readmem(TEMP_OUT,2,_buff);
+    readmem(TEMP_OUT, 2, _buff);
     *_Temp = 35 + (((_buff[0] << 8) | _buff[1]) + 13200) / 280.0;    // F=C*9/5+32
 }
 
 
 void
-ITG3200::readGyroRaw(int *_GyroX, int *_GyroY, int *_GyroZ){
+ITG3200::readGyroRaw(int16_t *_GyroX, int16_t *_GyroY, int16_t *_GyroZ){
     readmem(GYRO_XOUT, 6, _buff);
     *_GyroX = ((_buff[0] << 8) | _buff[1]);
     *_GyroY = ((_buff[2] << 8) | _buff[3]);
@@ -250,7 +246,7 @@ ITG3200::readGyroRaw(int *_GyroX, int *_GyroY, int *_GyroZ){
 
 
 void
-ITG3200::readGyroRaw(int *_GyroXYZ){
+ITG3200::readGyroRaw(int16_t *_GyroXYZ){
     readGyroRaw(_GyroXYZ, _GyroXYZ+1, _GyroXYZ+2);
 }
 
@@ -272,7 +268,7 @@ ITG3200::setGains(float _Xgain, float _Ygain, float _Zgain) {
 
 
 void
-ITG3200::setOffsets(int _Xoffset, int _Yoffset, int _Zoffset) {
+ITG3200::setOffsets(int16_t _Xoffset, int16_t _Yoffset, int16_t _Zoffset) {
     offsets[0] = _Xoffset;
     offsets[1] = _Yoffset;
     offsets[2] = _Zoffset;
@@ -280,11 +276,11 @@ ITG3200::setOffsets(int _Xoffset, int _Yoffset, int _Zoffset) {
 
 
 void
-ITG3200::zeroCalibrate(unsigned int totSamples, unsigned int sampleDelayMS) {
-    int xyz[3];
-    float tmpOffsets[] = {0,0,0};
+ITG3200::zeroCalibrate(uint16_t totSamples, uint16_t sampleDelayMS) {
+    int16_t xyz[3];
+    float tmpOffsets[] = {0, 0, 0};
 
-    for(uint i=0; i<totSamples; i++) {
+    for(uint16_t i=0; i<totSamples; i++) {
         QThread::msleep(sampleDelayMS);
         readGyroRaw(xyz);
         tmpOffsets[0] += xyz[0];
@@ -296,7 +292,7 @@ ITG3200::zeroCalibrate(unsigned int totSamples, unsigned int sampleDelayMS) {
 
 
 void
-ITG3200::readGyroRawCal(int *_GyroX, int *_GyroY, int *_GyroZ) {
+ITG3200::readGyroRawCal(int16_t *_GyroX, int16_t *_GyroY, int16_t *_GyroZ) {
     readGyroRaw(_GyroX, _GyroY, _GyroZ);
     *_GyroX += offsets[0];
     *_GyroY += offsets[1];
@@ -305,14 +301,14 @@ ITG3200::readGyroRawCal(int *_GyroX, int *_GyroY, int *_GyroZ) {
 
 
 void
-ITG3200::readGyroRawCal(int *_GyroXYZ) {
+ITG3200::readGyroRawCal(int16_t *_GyroXYZ) {
     readGyroRawCal(_GyroXYZ, _GyroXYZ+1, _GyroXYZ+2);
 }
 
 
 void
 ITG3200::readGyro(float *_GyroX, float *_GyroY, float *_GyroZ){
-    int x, y, z;
+    int16_t x, y, z;
 
     readGyroRawCal(&x, &y, &z); // x,y,z will contain calibrated integer values from the sensor
     *_GyroX =  x / 14.375 * polarities[0] * gains[0];
@@ -461,18 +457,5 @@ ITG3200::readmem(uint8_t _addr, uint8_t _nbytes, uint8_t __buff[]) {
         exit(-1);
     }
     close(fd);
-
-//    Wire.beginTransmission(_dev_address); // start transmission to device
-//    Wire.write(_addr); // sends register address to read from
-//    Wire.endTransmission(); // end transmission
-
-//    Wire.beginTransmission(_dev_address); // start transmission to device
-//    Wire.requestFrom(_dev_address, _nbytes);// send data n-bytes read
-//    uint8_t i = 0;
-//    while (Wire.available()) {
-//        __buff[i] = Wire.read(); // receive DATA
-//        i++;
-//    }
-//    Wire.endTransmission(); // end transmission
 }
